fix(neuron): seed GatedLIFNode spike state with the input dtype
multi_step_functional_forward() created the previous-spike state with
torch.zeros(..., device=x_seq.device) only, so it always defaulted to
float32 and promoted the membrane voltage and every later spike in a
half-precision GatedLIFNode.

+++ spikingjelly/activation_based/neuron/lif_variants.py
@@ -234,7 +234,7 @@ class GatedLIFNode(base.MemoryModule):
         linear_decay = self.linear_decay.view(1, -1, 1, 1).sigmoid()
         v_subreset = self.v_subreset.view(1, -1, 1, 1).sigmoid()
 
-        spike = torch.zeros(x_seq.shape[1:], device=x_seq.device)
+        spike = torch.zeros(x_seq.shape[1:], device=x_seq.device, dtype=x_seq.dtype)
         spike_seq = []
         for t in range(self.T):
             spike, v = functional.gated_lif_step(
